+++ client/web/divkit/typings/server.d.ts
@@ -1,9 +1,15 @@
-import type { DivJson, ErrorCallback, Platform } from './common';
+import type {
+    DivJson,
+    ErrorCallback,
+    Platform,
+    Customization
+} from './common';
 
 export function render(opts: {
     json: DivJson;
     id: string;
     platform?: Platform;
     mix?: string;
+    customization?: Customization;
     onError?: ErrorCallback;
 }): string;