+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 # Web Client:
 Added:
 * `mix` property for a custom class name
+* `customization` render property for a desktop `gallery`
 
 Fixed:
 * Incorrect `container` spacing with `separators`